+++ client/src/game/ui/settings/shape/TrackerSettings.vue
@@ -37,14 +37,11 @@ function toggleCompositeTracker(trackerId: string): void {
     activeShapeStore.removeTracker(trackerId, SyncTo.SHAPE);
 
     const oldShape = tracker.shape;
-    if (oldShape === activeShapeStore.state.uuid) {
-        activeShapeStore.setTrackerShape(tracker.uuid, activeShapeStore.state.parentUuid!);
-    } else {
-        activeShapeStore.setTrackerShape(tracker.uuid, activeShapeStore.state.uuid!);
-    }
+    const newShape =
+        oldShape === activeShapeStore.state.uuid ? activeShapeStore.state.parentUuid! : activeShapeStore.state.uuid!;
 
-    activeShapeStore.pushTracker(tracker, tracker.shape, SyncTo.SHAPE);
-    sendShapeMoveTracker({ shape: oldShape, new_shape: tracker.shape, tracker: tracker.uuid });
+    activeShapeStore.pushTracker(tracker, newShape, SyncTo.SHAPE);
+    sendShapeMoveTracker({ shape: oldShape, new_shape: newShape, tracker: tracker.uuid });
 }
 
 // Aura
@@ -72,14 +69,11 @@ function toggleCompositeAura(auraId: string): void {
     activeShapeStore.removeAura(auraId, SyncTo.SHAPE);
 
     const oldShape = aura.shape;
-    if (oldShape === activeShapeStore.state.uuid) {
-        activeShapeStore.setAuraShape(aura.uuid, activeShapeStore.state.parentUuid!);
-    } else {
-        activeShapeStore.setAuraShape(aura.uuid, activeShapeStore.state.uuid!);
-    }
+    const newShape =
+        oldShape === activeShapeStore.state.uuid ? activeShapeStore.state.parentUuid! : activeShapeStore.state.uuid!;
 
-    activeShapeStore.pushAura(aura, aura.shape, SyncTo.SHAPE);
-    sendShapeMoveAura({ shape: oldShape, new_shape: aura.shape, aura: aura.uuid });
+    activeShapeStore.pushAura(aura, newShape, SyncTo.SHAPE);
+    sendShapeMoveAura({ shape: oldShape, new_shape: newShape, aura: aura.uuid });
 }
 </script>
 
@@ -144,6 +138,7 @@ function toggleCompositeAura(auraId: string): void {
                     <input
                         v-show="isComposite"
                         type="checkbox"
+                        :checked="tracker.shape === activeShapeStore.state.parentUuid"
                         @click="toggleCompositeTracker(tracker.uuid)"
                         :disabled="!owned"
                         :title="t('common.toggle_public_private')"